Refactored some of the generic collision code into its own file.
